requirements.txt
# Основное
Django==5.2.5
django-environ==0.11.2
Pillow==10.4.0
whitenoise==6.7.0

# Для работы с формами и тегами (если используешь)
django-taggit==5.0.1

# Утилиты
sqlparse==0.5.1
asgiref==3.8.1

# Для деплоя (обычно нужны)
gunicorn==23.0.0
